+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2025-09-12
+- Added python-solution for task#3
+- [New task #3](./tasks.md#3-расчёт-стоимости-основных-производственных-фондов)
 ## 2025-09-11
 - Added 1-2 tasks for all languages
 - Init repo